refactor(sections): extract aligned spelling columns into spelling.rs
Pure motion: the reader for a row whose second column is another spelling
of the same option rather than the start of its description.
